Add files via upload
This is the beginning of a more advanced  program that will take a racer's name, assign them a number, and finally assign a race time. Based on the racer's age, a post-race location will be assigned. So far this program uses basic variable declarations,  an array, if/else statements, and a 'for' loop.  I am interested in how race websites generate race numbers and I would like to create my own program to execute some similar functions.